move global styles to styles dir
