+++ src/app/styles/GlobalStyles.tsx
